+++ AionLightning.NET/AionLightning.Game/Network/Aion/ServerPackets/SM_INVENTORY_INFO.cs
@@ -57,7 +57,7 @@ public sealed class SM_INVENTORY_INFO : AionServerPacket
         w.WriteH(0);
         w.WriteD(0);
 
-        w.WriteH(item.Slot);
+        w.WriteH((int)item.Slot);
         w.WriteC(0); // cloth flag
 
         if (item.EnchantLevel > 0)
